app/requirements.txt
gunicorn
flask
duo_client
slack_sdk
requests
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
